Refactor : 코드 리뷰에 따른 코드 수정
- 코드 리뷰에 따른 코드 수정

Ref: #140

+++ src/main/java/com/example/baseballprediction/domain/chat/minigame/entity/MiniGame.java
@@ -24,23 +24,23 @@ public class MiniGame {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "mini_game_id")
-	private int id;
+	private Long id;
 	
 	@ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "creator_member_id")
-    private Member creator;
+	@JoinColumn(name = "creator_member_id")
+	private Member creator;
 	
 	private String question;
-    private String option1;
-    private String option2;
-    
-    
-    @Builder
-    public MiniGame(Member creator,String question,String option1,String option2) {
-    	this.creator = creator;
-    	this.question = question;
-    	this.option1 = option1;
-    	this.option2 = option2;
-    }
+	private String option1;
+	private String option2;
+	
+	
+	@Builder
+	public MiniGame(Member creator,String question,String option1,String option2) {
+		this.creator = creator;
+		this.question = question;
+		this.option1 = option1;
+		this.option2 = option2;
+	}
 
 }
